Updated ft_strlen.c & libft.h

+++ 00_Libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: lyao-che <lyao-che@student.42kl.edu.my>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/23 11:02:58 by lyao-che          #+#    #+#             */
-/*   Updated: 2022/05/23 11:08:08 by lyao-che         ###   ########.fr       */
+/*   Updated: 2022/05/25 09:37:50 by lyao-che         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
-	int	counter;
+	size_t	counter;
 
 	counter = 0;
 	while(str[counter] != '\0')
